+++ course_1_homework/math_function.h
@@ -3,6 +3,6 @@
 #define FIRST_PROJECT_MATH_FUNCTION_H_
 
 int euclide_recursiv(int a, int b);
-int factorial_recursiv(int a);
+long long int factorial_recursiv(int a);
 
 #endif //FIRST_PROJECT_MATH_FUNCTION_H_
